Add job definitions for running Pulp Smash
Add job definition to run Pulp Smash against Fedora 22 and 23 and RHEL 6
and 7.

In order to have a very similar environment to a real use of Pulp the
jobs are designed to install Pulp on one machine and run Pulp Smash from
another machine. Translating to Jenkins, Pulp is installed on one slave
and Pulp Smash is run from other slave, at the end the jUnit report will
be published.