Improve pulp-dev jobs
Pulp Smash is able to ssh into pulp server and run some cli commands,
but, in order to do that, it needs a password-less ssh login setup to
the pulp server. Improve the pulp-dev jobs to setup password-less ssh
login on the pulp server and pass the private key to the
pulp-smash-runner job in order to proper ssh into the pulp server
machine.