Skip coverage report on RHEL6
The Python version available on RHEL6 is Python 2.6 which is a really old
version and the coverage report doesn't run on that version. Is better to skip
capturing the coverage report on that OS version in order to avoid the build to
fail and not publish the jUnit XML report.