Add nodepool bootstrap.sh setup script
The script will bootstrap a nodepool node in order to be ready to become a
Jenkins executor.

Also ignore files used for provision that may vary from deploy to deploy and
update the README to match the new requirements.

Finally update prepare_node.sh script to source the bootstrap script and use
its exported variables.