Update Pulp Automation jobs
Remove the need of an additional pool for atomic on the Pulp Ansible Role.

Because the update of the Pulp Ansible Role, update the pulp automation jobs to
pass the expected information to the playbook.

Finally update the unsubscribe procedure for RHEL based machines in order to
make it really run.