Adds a docs builder Jenkins job
This job nightly build one sphinx project which includes
platform and all plugins develped by the core dev team.

This job uses the docs-ownership so it has owners. It also
uses the e-mail notify macro for owner notification on
failure.

The version of each plugin is specified in the releases yaml
files used by the build infrastructure. Each repo is cloned
freshly with each build.

This job comes with additional rst files which are merged in
prior to the build. This additional content organizes all of
the other content which was not designed to be shown as one
site.

It also deploys a robots.txt for the root of the docs site.

https://pulp.plan.io/issues/950
re #950