Add Jenkins jobs' ownership and email notification
Create ownership macros and apply to each managed job, also add
email-notify-owners to each job that does not have an email sending publisher
configured in order to notify owners about failures and the job not being run.

Also add a job template with the minimal set of definitions and information on
README on how to use the template.