Improve pulp upgrade automation
Add jobs to test from current Y stream stable build to the current Y nightly
build. Make the job template generic in order to be able to upgrade from
any build (beta, nightly and stable) to any build by updating the repository
properly.

Add permutations that specify not just the Pulp version but the build as well,
this is going to allow more permutations and have more control and information
from which Pulp version and build are being upgrading from and to.

Fix missing `sudo` when restarting services after the upgrade.