fix deprecations and cobertura type problems
The cobertura plugin wants ints, not floats, for coverage values. This was
causing the "Error in request. Possibly authentication failed [500]: Server Error"
errors that we've been seeing.

groovy-postbuild was changed to want the groovy script in a "script"
dict config key, so I did that where needed to make that deprecation
warning go away

ssh-agent-credentials wants an array of users instead of a single user
param now in its config dict, so I've updated jobs using that wrapper
to be configured the new way. This has also become a macro, because
if I'm going to fix it the same way in four places I might as well
make it a macro.