Correctly compare Pulp version strings
When comparing two version strings, it's invalid to do this:

    version | float >= 2.8

The problem is that `version` may be a value such as '2.10', which is
cast to the float value of 2.1. The `version_compare` filter should be
used instead.

Thanks to @elyezer for help with tracking down this issue.

Fix: https://pulp.plan.io/issues/2080