Refactor satellite6-upgrade-pulp job
Convert the job to job template in order to reuse it for testing upstream and
downstream Satellite 6 with an upgraded version of Pulp. Also clean it up.

Create scripts directory to add the scrips included by the
satellite6-upgrade-pulp job.

Update the satellite-jenkins-trigger. Make it more generic by expecting the job
name and the job parameters.