Use updated RHEL images, remove grub hack
When using the "updated" rhel 7.3 image base, the el7 grub hack in
prepare_node is no longer necessary. This hack was causing a syntax
error in the grub config which prevented our prepared images from
booting.

I did test building the el6 images on the "updated" base, and they
appear to be working correctly.