Improved update-version-and-merge-forward
This was done in support of an odd and somewhat inappropriate usage of
this script, which is that it is the easiest way to `git clone` all of
the repos defined in a release config to the correct branch and ensure
the versions in that repo are all updated to reflect the release config
without taking any further action on that repo.

In particular, this is handy for the docs-builder workflow, which needs
to do all of the aforementioned things.

This script now:
- No longer explodes if asked to merge-forward a tag; it now refuses and
  moves on.
- While I was in there, I also added support for similarly skipping release
  branches, since the versions on those should never need to be updated by
  any process other than a merge.
- Flaked some 8s.