Pull the current pulp version from setup.py in pr tests
We currently pull this from a file that tito drops into the source repo,
which confuses the PR tester on master because this file only gets updated
by tito when we do a release build. The spec file and setup.pys do get
updated regularly, though. Since the spec files are moving to the packaging
repo, it makes sense to ask the setup.py file what the current version is
and use that, rather than depend on a file made by tito after a release.