Build scripts filter out unsupported dists by default
With the changes to our support policy, it's become necessary for
release engineering sanity to bake the idea of supported distributions
into the build scripts and related libraries.

Since we still want nightlies and the PR tester to give us feedback if
something we do needlessly breaks el6, the nightly build job has been
updated to include unsupported distributions when building packages.

This commit is also peppered with linting, as well as more preparatory
work getting us ready to switch to python3 when building pulp (mainly
because we can, and the work is largely already done).

closes #2576
https://pulp.plan.io/issues/2576