Use firewalld instead of iptables on most systems
The primary IPv4 firewall configuration tool on RHEL 6 is iptables. The
primary IPv4 firewall configuration tool on RHEL 7, Fedora 24 and Fedora
25 is firewalld. Update the Ansible roles for installing Pulp so that
firewalld is used in most cases, and add "python-firewalld" as a
dependency. The upshot is that firewall rules fore these distributions
are now persistent across system and service restarts.