Add support for unsupported package builds in plugins
In the first crack at making sure we only build supported packages
for correct dists, I made the incorrect assumption that only the
platform packages were affected. Puppet and RPM also need to be
supported to ensure we build the correct set of packages on
"unsupported" dists, like el5 and el6.

I tested this by running ./build-all 2.12-release with and without
the --build-unsupported option, observing that in each case the
puppet and rpm plugins were build for el5 and el6, and that the
correct set of packages are generated based on whether or not the
--build-unsupported flag was passed (they were).

closes #2679
https://pulp.plan.io/issues/2679