Move subscription-manager tasks into separate role
One of the roles called by the "pulp_server" playbook is the "pulp"
role. This role's job is to install and configure Pulp.  This role has
two sets of tasks: `tasks/main.yaml` and `tasks/pulp_server.yaml`.  The
former is almost entirely concerned with configuring
subscription-manager, and the latter does most of the actual work of
installing and configuring Pulp.

This seems a little strange: shouldn't the "pulp" role's `main.yaml`
install and configure Pulp? It's also a little problematic: there's no
easy way to skip the subscription-manager configuration, and this is
reflected by the large number of `when: ansible_distribution ==
'RedHat'` expressions in the `main.yaml` file.

Do the following:

* Make a new role named `subscription-manager`, and move the
  subscription-manager management logic into it.
* Move the firewall management code from the subscription-manager tasks
  to the pulp management tasks. After all, firewall management and
  configuration is needed to successfully get Pulp running, regardless
  of whether subscription-manager is present.
* Move the `pulp_install_prerequisites` variable. It controls whether
  the target system's firewall will be configured, and it seems like
  something that should always be done.
* Drop the `pulp_install_server` variable. If the Pulp server should be
  installed, that can now be accomplished by... running the `pulp` role.