Updates for python-redmine 2.0.0 release
The custom_fields.resources changed and needed to be updated. This
implementation is also more efficient because it fetches the field of
interest directly instead of having to iterate through a list to find
it.

I ran the job locally and it completed successfully. I saw it set some
needsinfos correctly, so I think this is working as expected.