Updates rebase-analysis and some ownership changes
* Updates for changes in python-redmine 2.0.0
* Some job ownership changes due to changing roles

The custom_fields.resources changed and needed to be updated. This
implementation is also more efficient because it fetches the field of
interest directly instead of having to iterate through a list to find
it.

I ran the rebase-analysis job locally and it completed successfully
with these chnages.

The docs-owernship, dev-ownership, and bz-redmine-ownership jobs had
their owners/co-owners changed some based on changing roles.