pulp-crane role: Let Apache bind to port 5000
Add a task to the pulp-crane role. This new task configures SELinux so
that Apache can bind to port 5000. This new task is a hack, and it
exists only because https://pulp.plan.io/issues/2719 has not yet been
fixed. When that issue is fixed, the hack added by this commit should be
removed.