Add roles to populate, backup, and restore pulp
Adds two playbooks and several new roles.

The `pulp_backup.yaml` playbook uses existing roles to install pulp on a
host, and then uses the new roles pulp-populate and pulp-backup to
populate it with content and back it up. Optionally, the user can fetch
the backup files to the control node.

The `pulp_restore.yaml` playbook uses existing roles to install pulp on
a host and uses the new role pulp-restore-backup to restore the files
and (optionally) test whether or not the restore succeeded.

Notes for how to use these playbooks are found in the playbooks
themselves.