Add plugins to the Pulp docs according to the release config
For any component 'build_docs' option can be specified
in the release config. If disabled, it allows not to build docs
for plugins if not needed (e.g. community plugin hosts their
docs somewhere else).

Index files for platform docs are moving to the platform repository.
Plugins' index file will be copyied in as before but unnecessary
plugins will be filtered out from it.