Unregister jenkins nodes as a post-build action
Previously, if a build on a RHEL system failed after registering, but
before unregistering, the only recouse left to release the entitlements
the system used up would be through the customer portal.

This PR adds the unregister step as a post-build action that happens no
matter if the build fails or succeeds.