Make pulp-3-dev-{os} jobs smash Pulp 3
Of note is that the pulp-smash-runner job is not being called. Instead,
a few extra shell commands are being executed.

This approach definitely has some downsides. Among other things, the
existing pulp-smash-runner job is highly configurable, where a user can
set options such as the broker used by the Pulp installation under test.
(e.g. qpid or rabbitmq) However, the job is also very complicated,
making calls to it much harder to accomplish. And the job performs some
tasks that are specific to Pulp 2, such as working with Pulp 2's SSL
certificates.

In this case, it seems better to have a good job now rather than a
perfect job later. This is doubly true if Jenkins 2 will soon be
available, which will make things like pipelines available.