Revert "Require Fedora 26 or newer for pulp-fixtures"
This reverts commit 5d997ebbcc413f6fc7f395086612967c31b50e0d. This
commit let the pulp-fixtures-publisher job depend on 'f26-np || f27-np'.
While this change does update bash, it breaks pulp-fixtures-publisher,
because hosts with this label don't have Docker installed and
configured, and pulp-fixtures also depends on Docker. This could be
solved by installing and configuring Docker on nodes with these labels.
A lower-impact solution is to revert the commit above, and let docker-np
hosts inherit from a Fedora 26 base image.