Add FIPS installation role
Add a new role called pulp-fips, and call it from the `pulp_server.yaml`
playbook. This new role is only being used for RHEL 7, but it should be
applicable to other operating systems such as CentOS 7. This new role
only runs when enable_fips=True is passed in the anisble-playbook
command.