Adding F28 python3 firewalld bindings
This PR addresses two issues that are related to Fedora28.

It makes sure Python3 interpreter is used for running Ansible in Fedora
28. This was required since the python2-firewall packages are not used
in F28.
SELinux is set in the permissive mode in F28, Since the tasks that
starts squid.service [1] fails in F28.

Since the travis builds fails for testing jjb. Updating JJB in the
travis jobs.

[1] https://pulp.plan.io/issues/3926