Migrating ``disk-image-builder-base`` job to Open Stack node.
The current ``disk-image-builder-base`` job is running on a physical
instance. Migrating this job to be run on open stack gives much more
availability. As a part of this commit, a new instance called
``disk-image-builder-os`` is set up in the jenkins-openstack cloud
plugin which points to a fedora 27 instance in openstack. Once this job
runs, the fedora-27 instance gets provisioned which will act as a host
for disk image building.