Adding rhel-os and rhel-np images for upgrade and unittest jobs
The current rhel-os label was pointing to a nodepool node. This was
done since the rhel-os configured with Disk-Image-Builder is
incompatible with the `unit-test` jobs that requires older builds of
certain packages. However the nodepool nodes were making the upgrade job
to fail due to java installation. Thus this commit takes care of the
following

* Adding a new label rhel7-np for the Unit-Test jobs that runs on the
old nodepool built rhel7
* Configuring all the other jobs to run on the rhel7 build by
Disk-Image-Builder

Once this PR is merged, Will configure
* rhel7-os
* rhel7-np

nodes inside jenkins to point to the appropriate machines.