Adding Safe delete logic to disk image builder
The current implementation of Disk Image Builder deletes an image in the
upstream without considering the status of the newly created image. Thus
even before the new image created is uploaded, the old image is deleted
. This was done to remove any name conflicts, however this creates a
threat of removing a working image in the case of failure. This commit
takes care of preserving the old image when the latest build fails.