Pin pytest version for Pulp 2 tests
Latest version of pytest `4.2.0` in somehow changed the order of
tests running - mainly related to inheritance. Tests inherinted from the
``BaseAPICrudTestCase`` are running the  ``BaseAPICrudTestCase`` as a
test itself. This behaviour is not observed in older versions of pytest.

Pin pytest version until this behaviour is figured out.

See: https://github.com/pytest-dev/pytest/releases