Changing Ansible Installer priorities with current 2.18.1 definition
To solve an EPEL version problem with the python-celery stack, the
final solution during build has changed the EPOCH on NEVRA for those RPMs
to 10:

The resulting files needed to be updated to:
* use priority on FIPS with EPEL
* do not use priority on non-FIPS with EPEL

The end state of each of those upgrade permutations leave the box in the same
RPM state without the end-user to have to play with yum or repository
priorities.

Removing the pulp-priority role as it is now deprecated with these changes.

The included, tested changes are the output of the solution for CI.

refs #4387