Update pulp role to define defaults for repo_priorty and evalute run-time
As an addendum to PR #644 , the repo_priorty variable in the pulp role
must be defined within the scope of the role (not the pulp_server
playbook) as other playbooks use that role.

Making changes to:
- Declare a default of '99', which is the same as no priority
- Declare a fips priority of '55' for fips pulp installs
- Evaluate the proper priority within the pulp role for the pulp.repos from fips
- Remove the pre_task declaration from the pulp_server playbook

Retested on fips and non-fips with the pulp_server playbook as well
as other playbooks dependent on the pulp role.

refs #4359