Increase pulp2 versioning informaton for audit
During any test run, there may be a question about the content of the
install of an install.

Adding the full information for ``pulp-admin status`` and all pulp
installed RPMs to compare against repos.fedorapeople.org. This should
allow an auditor to use the result to compare instead of having to spawn
a new machine to then compare the running version. That may not be possible
at times as the testing repos are not under QE control.