Updates to add hostname to /etc/hosts since DNS is unreliable
PnT DNS for our CI network is not reliable or not returning the
fully qualified hostname from DNS.

Updating required jobs to pass the IP and HOSTNAME to pulp-smash and
pulp-smash to expect those parameters to add to /etc/hosts.

If these parameters are NOT defined, a blank line is added to /etc/hosts.

If DNS IS working, the previous hostname definition should continue to work
allowing for previously functioning behavior.