Fixing attrs to version 19.1.0 to avoid pyest 'convert' issues
With an update to attrs==19.2.0, a conflict was noted in pytest with an error
about "keyword arguement 'convert'.

Reinstalling version 19.1.0 to avoid this problem in the short-term and allow
ci to start functioning again.

A long-term solution relative to pytest==4.1.0 pinning needs to be resolved,
which is the crux of the issue.

Please see reference JIRA issue for more background.

Ref: SATQE-5885