Lock users instead of deleting them
In addition to that, this commit contains patches which resolve the recent job failures. For example, some UI elements were not loaded before performing the click action. Because of that, it was not possible to delete spam content and the job failed with runtime exceptions.

Also, when sending batch update requests, timeout errors were raised. This could be caused by the size and format of sent requests. The script was trying to delete almost 600 rows where every single row was considered to be a separate update to the spreadsheet.